# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Boisolo Api",
        "description": "This is a sample API documentation",
        "contact": {
            "email": "dev@boisolo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/order-stats/last": {
            "post": {
                "tags": [
                    "OrderStats"
                ],
                "summary": "Obtener la última estadística de un pedido",
                "description": "Retorna la última estadística de pedidos para una línea de producción usando un token.",
                "operationId": "ee47f14473643b8ecb7469ddd308a775",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "some-production-line-token"
                                    },
                                    "production_line_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "order_id": {
                                        "type": "string",
                                        "example": "ORD12345"
                                    },
                                    "units": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "box": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "units_box": {
                                        "type": "integer",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Última estadística de pedido obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "production_line_name": {
                                            "type": "string",
                                            "example": "Linea 1"
                                        },
                                        "order_id": {
                                            "type": "string",
                                            "example": "ORD12345"
                                        },
                                        "units": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-10-15 14:35:12"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Token es requerido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Token is required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Línea de producción no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Production line not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/order-stats/between-dates": {
            "post": {
                "tags": [
                    "OrderStats"
                ],
                "summary": "Obtener estadísticas de pedidos entre fechas",
                "description": "Obtiene estadísticas de pedidos entre dos fechas dadas para una línea de producción específica, utilizando el token de la línea.",
                "operationId": "89d56fec7821f1729d93f47db5a82eee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "example": "some-production-line-token"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "example": "2024-10-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "example": "2024-10-20"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estadísticas de pedidos entre fechas obtenidas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "production_line_name": {
                                            "type": "string",
                                            "example": "Linea 1"
                                        },
                                        "order_id": {
                                            "type": "string",
                                            "example": "ORD12345"
                                        },
                                        "units": {
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-10-15 14:35:12"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Línea de producción no encontrada o sin estadísticas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Production line not found or no order stats found for the specified dates"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}